LPS-78208 Move JournalItemSelectorWebKeys to internal.constants
