+++ modules/apps/web-experience/journal/journal-item-selector-web/src/main/java/com/liferay/journal/item/selector/web/internal/constants/JournalItemSelectorWebKeys.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.journal.item.selector.web.internal;
+package com.liferay.journal.item.selector.web.internal.constants;
 
 /**
  * @author Roberto Díaz
