+++ modules/apps/web-experience/journal/journal-item-selector-web/src/main/resources/META-INF/resources/init.jsp
@@ -22,7 +22,7 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 
 <%@ page import="com.liferay.document.library.display.context.DLMimeTypeDisplayContext" %><%@
 page import="com.liferay.document.library.kernel.util.DLUtil" %><%@
-page import="com.liferay.journal.item.selector.web.internal.JournalItemSelectorWebKeys" %><%@
+page import="com.liferay.journal.item.selector.web.internal.constants.JournalItemSelectorWebKeys" %><%@
 page import="com.liferay.journal.item.selector.web.internal.context.JournalItemSelectorViewDisplayContext" %><%@
 page import="com.liferay.journal.model.JournalArticle" %><%@
 page import="com.liferay.portal.kernel.dao.search.SearchContainer" %><%@
